Fixed a logic error.
When creating a filestore, the root of the filestore will be the
provided path.

For instance, say you have a git repository at /var/repo1 that contains
directories dir1 and dir2.
If you create a git filestore with

   repo <- gitFileStore "/var/repo1/dir1"

The root path for the filestore will be dir1 and not repo1.
I was not aware of that and therefore we committed files relative to the
filestore root rather than git root.